(spas:feature:layout) renamed files
