+++ src/spas/apps/jwadmin/src/app/app.component.html
@@ -1,4 +1,2 @@
-<!-- <ui-toolbar isNavigationEnabled="true" [isTopNavigation]="isTopNavigation$ | async"></ui-toolbar> -->
-<!-- <feature-layout-toolbar-default></feature-layout-toolbar-default> -->
-<feature-layout></feature-layout>
+<layout-feature></layout-feature>
 <member-management></member-management>
