+++ src/spas/apps/jwadmin/src/app/app.module.ts
@@ -3,7 +3,6 @@ import { NgModule } from '@angular/core';
 import { AngularFireModule } from '@angular/fire';
 
 import { AppComponent } from './app.component';
-import { SharedUiLayoutModule } from '@nimb3s/shared/ui/layout';
 import { LayoutFeatureModule } from '@nimb3s/layout/feature';
 import { environment } from '../environments/environment';
 import { MemberManagementFeatureModule } from '@nimb3s/member-management/feature'
@@ -13,7 +12,6 @@ import { MemberManagementFeatureModule } from '@nimb3s/member-management/feature
   imports: [
     BrowserModule,
     AngularFireModule.initializeApp(environment.firebase, 'dev-jwadmin'),
-    SharedUiLayoutModule,
     LayoutFeatureModule,
     MemberManagementFeatureModule
   ],
