+++ src/spas/libs/layout/feature/src/index.ts
@@ -1,1 +1,1 @@
-export * from './lib/feature-layout.module';
+export * from './lib/layout-feature.module';
