+++ src/spas/libs/layout/feature/src/lib/layout-feature.module.spec.ts
@@ -1,5 +1,5 @@
 import { async, TestBed } from '@angular/core/testing';
-import { LayoutFeatureModule } from './feature-layout.module';
+import { LayoutFeatureModule } from './layout-feature.module';
 
 describe('LayoutFeatureModule', () => {
   beforeEach(async(() => {
