+++ src/spas/libs/layout/feature/tslint.json
@@ -1,7 +1,7 @@
 {
   "extends": "../../../tslint.json",
   "rules": {
-    "directive-selector": [true, "attribute", "feature-layout", "camelCase"],
-    "component-selector": [true, "element", "feature-layout", "kebab-case"]
+    "directive-selector": [true, "attribute", "layout-feature", "camelCase"],
+    "component-selector": [true, "element", "layout-feature", "kebab-case"]
   }
 }
